# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recovery Velocity Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parishes/{parishId}/resources:
    get:
      operationId: getParishResources
      tags: [parishes]
      summary: Resource status indicators for a parish (fuel, water, medical)
      parameters:
        - in: path
          name: parishId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParishResources"
        "404":
          description: Not found
  /parishes/{parishId}/infrastructure:
    get:
      operationId: getParishInfrastructure
      tags: [parishes]
      summary: List parish infrastructure references and partner-reported availability
      parameters:
        - in: path
          name: parishId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parish infrastructure sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureOverview"
        "404":
          description: Parish not found
  /parishes/{parishId}/infrastructure/{category}:
    get:
      operationId: getParishInfrastructurePlaces
      tags: [parishes]
      summary: Load public Google place listings for a parish infrastructure category
      parameters:
        - in: path
          name: parishId
          required: true
          schema:
            type: string
        - in: path
          name: category
          required: true
          schema:
            $ref: "#/components/schemas/InfrastructureCategory"
      responses:
        "200":
          description: Parish infrastructure sources with Google place listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfrastructureOverview"
        "400":
          description: Unknown infrastructure category
        "404":
          description: Parish not found
...
components:
  schemas:
...
    ParishResources:
      type: object
      required: [parishId, fuelPercent, waterPercent, medicalPercent, lastUpdated]
      properties:
        parishId:
          type: string
        fuelPercent:
          type: number
        waterPercent:
          type: number
        medicalPercent:
          type: number
        lastUpdated:
          type: string
    InfrastructureCategory:
      type: string
      enum: [hotel, bank, church_faith, fuel, medical, food, shelter, logistics, other]
    InfrastructureResourceOffer:
      type: object
      required: [resourceType, quantity, unit, availabilityWindow]
      properties:
        resourceType:
          type: string
        quantity:
          type: number
        unit:
          type: string
        availabilityWindow:
          type: string
          enum: [immediately, within_6h, within_24h, within_48h]
    InfrastructurePartnerOffer:
      type: object
      required: [id, organizationName, organizationType, parishId, facilityCategory, location, resources, source, verificationStatus, lastUpdated]
      properties:
        id:
          type: string
        organizationName:
          type: string
        organizationType:
          type: string
        parishId:
          type: string
        facilityCategory:
          $ref: "#/components/schemas/InfrastructureCategory"
        location:
          type: string
        contactNotes:
          type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/InfrastructureResourceOffer"
        source:
          type: string
          enum: [private_sector_submission]
        verificationStatus:
          type: string
          enum: [demo_seed, unverified_submission, partner_reported, verified]
        lastUpdated:
          type: string
    GoogleMapsSearchLink:
      type: object
      required: [category, label, url]
      properties:
        category:
          $ref: "#/components/schemas/InfrastructureCategory"
        label:
          type: string
        url:
          type: string
    GoogleMapsDirectory:
      type: object
      required: [status, sourceName, message, searchLinks, places]
      properties:
        status:
          type: string
          enum: [idle, available, unavailable, stale]
        sourceName:
          type: string
        message:
          type: string
        searchLinks:
          type: array
          items:
            $ref: "#/components/schemas/GoogleMapsSearchLink"
        places:
          type: array
          items:
            $ref: "#/components/schemas/GoogleMapsPlace"
        fetchedAt:
          type: string
        expiresAt:
          type: string
    GoogleMapsPlace:
      type: object
      required: [placeId, name, category, address, mapUrl]
      properties:
        placeId:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/InfrastructureCategory"
        address:
          type: string
        mapUrl:
          type: string
        rating:
          type: number
        ratingCount:
          type: number
        phoneNumber:
          type: string
        websiteUrl:
          type: string
        types:
          type: array
          items:
            type: string
        openNow:
          type: boolean
    InfrastructureOverview:
      type: object
      required: [parishId, parishName, categories, googleDirectory, partnerOffers, lastUpdated]
      properties:
        parishId:
          type: string
        parishName:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/GoogleMapsSearchLink"
        googleDirectory:
          $ref: "#/components/schemas/GoogleMapsDirectory"
        partnerOffers:
          type: array
          items:
            $ref: "#/components/schemas/InfrastructurePartnerOffer"
        lastUpdated:
          type: string
...