# lib/api-zod/src/generated/types/infrastructureResourceOffer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Recovery Velocity Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InfrastructureResourceOfferAvailabilityWindow } from './infrastructureResourceOfferAvailabilityWindow';

export interface InfrastructureResourceOffer {
  resourceType: string;
  quantity: number;
  unit: string;
  availabilityWindow: InfrastructureResourceOfferAvailabilityWindow;
}