# lib/api-zod/src/generated/types/parishResources.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Recovery Velocity Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ParishResources {
  parishId: string;
  fuelPercent: number;
  waterPercent: number;
  medicalPercent: number;
  lastUpdated: string;
}